fix: ensure filtered image persists through crash recovery
